+++ modelcraft/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.0.0"
+__version__ = "4.0.0"
 
 from .cell import max_distortion as max_cell_distortion
 from .cell import remove_scale
@@ -11,16 +11,19 @@ from .jobs.buccaneer import Buccaneer
 from .jobs.comit import Comit
 from .jobs.coot import FixSideChains
 from .jobs.coot import Prune
-from .jobs.coot import RsrMorph
 from .jobs.ctruncate import CTruncate
+from .jobs.emda import EmdaMapMask
 from .jobs.findwaters import FindWaters
 from .jobs.freerflag import FreeRFlag
 from .jobs.molrep import Molrep
 from .jobs.nautilus import Nautilus
 from .jobs.parrot import Parrot
 from .jobs.phasematch import PhaseMatch
-from .jobs.refmac import RefmacEm
-from .jobs.refmac import RefmacXray
+from .jobs.refmac import Refmac
+from .jobs.refmac import RefmacMapToMtz
+from .jobs.servalcat import ServalcatNemap
+from .jobs.servalcat import ServalcatRefine
+from .jobs.servalcat import ServalcatTrim
 from .jobs.sheetbend import Sheetbend
 from .pipeline import Pipeline
 from .reflections import DataItem
@@ -43,6 +46,8 @@ __all__ = [
     "contains_residue",
     "CTruncate",
     "DataItem",
+    "EmdaMapMask",
+    "entry_contents",
     "FindWaters",
     "FixSideChains",
     "FreeRFlag",
@@ -56,15 +61,17 @@ __all__ = [
     "PolymerType",
     "Prune",
     "read_structure",
-    "RefmacEm",
-    "RefmacXray",
+    "Refmac",
+    "RefmacMapToMtz",
     "remove_non_library_atoms",
     "remove_non_protein",
     "remove_residues",
     "remove_scale",
     "rmsz",
-    "RsrMorph",
     "run",
+    "ServalcatNemap",
+    "ServalcatRefine",
+    "ServalcatTrim",
     "Sheetbend",
     "update_cell",
     "write_mmcif",