+++ modelcraft/jobs/coot.py
@@ -20,7 +20,17 @@ class Coot(Job):
         self.fphis = fphis
 
     def _setup(self) -> None:
-        script_lines = ["try:\n", "    turn_off_backup(0)\n"]
+        script_lines = [
+            "try:\n",
+            "    COOT1 = True\n",
+            "    try:\n",
+            "        import coot_utils\n",
+            "    except NameError:\n",
+            "        COOT1 = False\n",
+            "    if COOT1:\n",
+            "        from coot import *\n",
+            "    turn_off_backup(0)\n",
+        ]
         for i, structure in enumerate(self.structures):
             write_mmcif(self._path(f"xyzin{i}.cif"), structure)
             script_lines += [
@@ -32,10 +42,11 @@ class Coot(Job):
                 f"    IMAP{i} = make_and_draw_map('hklin{i}.mtz', "
                 f"'{fphi.label(0)}', '{fphi.label(1)}', '', 0, 0)\n"
             ]
+        script_lines += ["    set_imol_refinement_map(IMAP0)\n"]
         for line in self.script.split("\n"):
             script_lines += [f"    {line}\n"]
         script_lines += [
-            "    write_cif_file(0, 'xyzout.cif')\n",
+            "    write_cif_file(IMOL0, 'xyzout.cif')\n",
             "    coot_real_exit(0)\n",
             "except:\n",
             "    import traceback\n",
@@ -91,12 +102,3 @@ class FixSideChains(Coot):
         super().__init__(
             script=script, structures=[structure], fphis=[fphi_best, fphi_diff]
         )
-
-
-class RsrMorph(Coot):
-    def __init__(self, structure: gemmi.Structure, fphi_best: DataItem):
-        path = os.path.join(os.path.dirname(__file__), "..", "coot", "morph.py")
-        with open(path) as stream:
-            script = stream.read()
-        script += "rsr_morph(IMOL0, IMAP0)\n"
-        super().__init__(script=script, structures=[structure], fphis=[fphi_best])