+++ modelcraft/monlib.py
@@ -53,3 +53,29 @@ def volume(code: str) -> float:
 @functools.lru_cache(maxsize=None)
 def weight(code: str) -> float:
     return sum(atom.el.weight for atom in chemcomp(code).atoms)
+
+
+@functools.lru_cache(maxsize=None)
+def group(code: str) -> gemmi.ChemComp.Group:
+    doc = gemmi.cif.read(_path(code))
+    monlib = gemmi.MonLib()
+    monlib.read_monomer_doc(doc)
+    return monlib.monomers[code].group
+
+
+@functools.lru_cache(maxsize=None)
+def is_protein(code: str) -> bool:
+    return group(code) in {
+        gemmi.ChemComp.Group.Peptide,
+        gemmi.ChemComp.Group.PPeptide,
+        gemmi.ChemComp.Group.MPeptide,
+    }
+
+
+@functools.lru_cache(maxsize=None)
+def is_nucleic(code: str) -> bool:
+    return group(code) in {
+        gemmi.ChemComp.Group.Dna,
+        gemmi.ChemComp.Group.Rna,
+        gemmi.ChemComp.Group.DnaRna,
+    }