+++ modelcraft/scripts/modelcraft.py
@@ -1,10 +1,16 @@
+import sys
+from ..arguments import parse
 from ..environ import setup_environ
-from ..modelcraft import ModelCraft
+from ..modelcraftem import ModelCraftEm
+from ..modelcraftxray import ModelCraftXray
 
 
 def main(args=None):
     setup_environ()
-    ModelCraft(args).run()
+    raw_args = args or sys.argv[1:]
+    parsed_args = parse(args)
+    pipeline = ModelCraftEm if parsed_args.mode == "em" else ModelCraftXray
+    pipeline(parsed_args, raw_args).run()
 
 
 if __name__ == "__main__":