+++ modelcraft/tests/ccp4/test_findwaters.py
@@ -1,3 +1,4 @@
+import gemmi
 from modelcraft.jobs.findwaters import FindWaters
 from modelcraft.structure import ModelStats
 from . import insulin_refmac
@@ -28,6 +29,10 @@ def test_insulin_dummy():
     assert stats_out.residues == stats_in.residues
     assert stats_out.waters == stats_in.waters
     assert stats_out.dummy_atoms > stats_in.dummy_atoms
+    for chain in findwaters.structure[0]:
+        for residue in chain:
+            if residue.name == "DUM":
+                assert residue[0].element == gemmi.Element("O")
 
 
 def test_existing_water_chain():
@@ -40,9 +45,6 @@ def test_existing_water_chain():
         structure=findwaters1.structure,
         fphi=refmac.fphi_best,
     ).run()
-    chains1 = len(findwaters1.structure[0])
-    chains2 = len(findwaters2.structure[0])
-    assert chains1 == chains2
     waters1 = ModelStats(findwaters1.structure).waters
     waters2 = ModelStats(findwaters2.structure).waters
     assert waters2 > waters1